refactor: consolidate auth under global package
