--- src/main/java/com/domisa/domisa_backend/auth/dto/LoginRequest.java
@@ -1,3 +1,0 @@
-package com.domisa.domisa_backend.auth.dto;
-
-public record LoginRequest(String authorizationCode) {}
