--- src/main/java/com/domisa/domisa_backend/auth/dto/LoginResponse.java
@@ -1,10 +1,0 @@
-package com.domisa.domisa_backend.auth.dto;
-
-public record LoginResponse(StatusDto status) {
-
-    public record StatusDto(
-            boolean isRegistered,
-            boolean hasIntroduction,
-            boolean isProfileCompleted
-    ) {}
-}
