--- src/main/java/com/domisa/domisa_backend/auth/handler/OAuth2LoginSuccessHandler.java
@@ -1,1 +1,0 @@
-// 이 파일은 더 이상 사용하지 않습니다. 삭제 예정.
