+++ src/main/java/com/domisa/domisa_backend/global/auth/dto/LoginRequest.java
@@ -1,0 +1,4 @@
+package com.domisa.domisa_backend.global.auth.dto;
+
+public record LoginRequest(String authorizationCode) {
+}
