+++ src/main/java/com/domisa/domisa_backend/global/auth/dto/LoginResponse.java
@@ -1,0 +1,11 @@
+package com.domisa.domisa_backend.global.auth.dto;
+
+public record LoginResponse(StatusDto status) {
+
+	public record StatusDto(
+		boolean isRegistered,
+		boolean hasIntroduction,
+		boolean isProfileCompleted
+	) {
+	}
+}
