+++ src/main/java/com/domisa/domisa_backend/global/auth/filter/JwtAuthenticationFilter.java
@@ -1,0 +1,53 @@
+package com.domisa.domisa_backend.global.auth.filter;
+
+import com.domisa.domisa_backend.global.auth.jwt.JwtProvider;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+public class JwtAuthenticationFilter extends OncePerRequestFilter {
+
+	private final JwtProvider jwtProvider;
+
+	public JwtAuthenticationFilter(JwtProvider jwtProvider) {
+		this.jwtProvider = jwtProvider;
+	}
+
+	@Override
+	protected void doFilterInternal(
+		HttpServletRequest request,
+		HttpServletResponse response,
+		FilterChain filterChain
+	) throws ServletException, IOException {
+
+		extractTokenFromCookie(request, "accessToken")
+			.filter(jwtProvider::validate)
+			.ifPresent(token -> {
+				Long userId = jwtProvider.getUserId(token);
+				UsernamePasswordAuthenticationToken auth =
+					new UsernamePasswordAuthenticationToken(userId, null, List.of());
+				SecurityContextHolder.getContext().setAuthentication(auth);
+			});
+
+		filterChain.doFilter(request, response);
+	}
+
+	private Optional<String> extractTokenFromCookie(HttpServletRequest request, String cookieName) {
+		if (request.getCookies() == null) {
+			return Optional.empty();
+		}
+		return Arrays.stream(request.getCookies())
+			.filter(cookie -> cookieName.equals(cookie.getName()))
+			.map(Cookie::getValue)
+			.findFirst();
+	}
+}
