+++ src/main/java/com/domisa/domisa_backend/global/auth/service/AuthService.java
@@ -1,0 +1,63 @@
+package com.domisa.domisa_backend.global.auth.service;
+
+import com.domisa.domisa_backend.domain.user.entity.User;
+import com.domisa.domisa_backend.domain.user.repository.UserRepository;
+import com.domisa.domisa_backend.global.auth.dto.LoginResponse;
+import com.domisa.domisa_backend.global.auth.jwt.JwtProvider;
+import com.domisa.domisa_backend.global.auth.oauth.KakaoOAuthService;
+import jakarta.servlet.http.HttpServletResponse;
+import java.time.Duration;
+import org.springframework.stereotype.Service;
+
+@Service
+public class AuthService {
+
+	private final KakaoOAuthService kakaoOAuthService;
+	private final UserRepository userRepository;
+	private final JwtProvider jwtProvider;
+	private final AuthCookieManager authCookieManager;
+
+	public AuthService(
+		KakaoOAuthService kakaoOAuthService,
+		UserRepository userRepository,
+		JwtProvider jwtProvider,
+		AuthCookieManager authCookieManager
+	) {
+		this.kakaoOAuthService = kakaoOAuthService;
+		this.userRepository = userRepository;
+		this.jwtProvider = jwtProvider;
+		this.authCookieManager = authCookieManager;
+	}
+
+	public LoginResponse login(String authorizationCode, HttpServletResponse response) {
+		String kakaoAccessToken = kakaoOAuthService.getAccessToken(authorizationCode);
+		String kakaoId = kakaoOAuthService.getKakaoId(kakaoAccessToken);
+
+		User user = userRepository.findByKakaoId(kakaoId)
+			.orElseGet(() -> userRepository.save(User.create(kakaoId)));
+
+		String accessToken = jwtProvider.createAccessToken(user.getId());
+		String refreshToken = jwtProvider.createRefreshToken(user.getId());
+
+		authCookieManager.addCookie(
+			response,
+			"accessToken",
+			accessToken,
+			Duration.ofMillis(jwtProvider.getAccessTokenValidityMs())
+		);
+		authCookieManager.addCookie(
+			response,
+			"refreshToken",
+			refreshToken,
+			Duration.ofMillis(jwtProvider.getRefreshTokenValidityMs())
+		);
+
+		return new LoginResponse(
+			new LoginResponse.StatusDto(
+				user.getIsRegistered(),
+				user.getHasIntroduction(),
+				user.getIsProfileCompleted()
+			)
+		);
+	}
+}
